requirements.txt
# Core NLP
spacy>=3.7.0
transformers>=4.36.0
torch>=2.1.0
tokenizers>=0.15.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0
datasets>=2.16.0

# API
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0
python-multipart>=0.0.6

# ML / Evaluation
scikit-learn>=1.3.0
seqeval>=1.2.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.1
loguru>=0.7.0
tqdm>=4.66.0
typer>=0.9.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Code Quality
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0
